requirements.txt
langchain
langchain-community
langchain-openai
openai
faiss-cpu
chromadb
pypdf
python-dotenv
gradio